BAD HACK: Set /statistics test to 0
Normally /statistics would return the number of sources contained in the
database.

The test (wrongly?) assumes that this will be always exactly 1, but when
starting with an empty database (and using the MongoDB instance given as
secret), the number of sources is, expectedly, zero. This hack forces the
test to check for 0 instead of 1, which will cause it to fail if the remote
database ever stores a source, but without the hack, the test would fail
anyway if the DB contained more than 1 sources.